Centralize pre persis/update temporal handlers.
Remove superfluous getters/setters for base class.
